Get Canadian data from the Borealis Repository This function is called by get_db_table if the public option is set to TRUE.

Get Canadian data from the Borealis Repository This function is called by get_db_table if the public option is set to TRUE.

Usage

get_canadian_data(
  table_name = NULL,
  dataverse_url = "https://borealisdata.ca",
  dataverse_doi = "doi:10.5683/SP3/QGLCKO",
  dataverse_api = getOption("drivesR.default.dataversetoken"),
  borealis_repo_info = NULL,
  directus_url = getOption("drivesR.default.url")
)

Arguments

table_name

Name of the drives database table, as shown in table_dictionary.

dataverse_url

Server URL for the Borealis repository, part of Dataverse. By default is set to "https://borealisdata.ca"

dataverse_doi

Persistent identifier for the collection in Borealis. This is set to a default.

dataverse_api

API key for the Borealis repository. This option is included for troubleshooting before the repository data tables are published. Once they are published, this option can remain as NULL.

borealis_repo_info

A dataframe containing Borealis file identifiers for each table_name. If NULL, this information is imported from Directus. Set with set_default_token().

directus_url

URL for the Directus database (can be set with global options).

Value

A dataframe with rows for the two Canadian sites for the specified table_name. This can be combined with the public Direcctus table using bind_rows.